Re-designed the payments system from scratch
