+++ src/main/java/com/dylankilbride/grouppay/Controllers/PaymentController.java
@@ -1,46 +1,31 @@
 package com.dylankilbride.grouppay.Controllers;
 
-import com.dylankilbride.grouppay.Clients.StripeClient;
 import com.dylankilbride.grouppay.Models.StripeCharge;
-import com.dylankilbride.grouppay.Models.StripeChargeReceipt;
+import com.dylankilbride.grouppay.Services.PaymentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/payments")
 public class PaymentController {
 
-	private StripeClient stripeClient;
-
 	@Autowired
-	public PaymentController(StripeClient stripeClient) {
-		this.stripeClient = stripeClient;
-	}
+	PaymentService paymentService;
 
-	@RequestMapping(value = "/charge",
+	@RequestMapping(value = "/oneTimePayment",
 					method = RequestMethod.POST,
-					produces = MediaType.APPLICATION_JSON_VALUE,
 					consumes = MediaType.APPLICATION_JSON_VALUE)
 	@ResponseBody
-	public StripeChargeReceipt chargeCard(@RequestBody StripeCharge stripeCharge) throws Exception {
-		return this.stripeClient.chargePaymentCardAndDontSave(stripeCharge.getTokenId(),
-						stripeCharge.getAmountInclFees(),
-						stripeCharge.getAmountForGroup(),
-						stripeCharge.getUserId(),
-						stripeCharge.getGroupAccountId());
+	public ResponseEntity makeOneTimePayment(@RequestBody StripeCharge chargeDetails) {
+		return paymentService.makeOneTimePaymentCharge(chargeDetails);
 	}
 
-	@RequestMapping(value = "/charge/save",
-					method = RequestMethod.POST,
-					produces = MediaType.APPLICATION_JSON_VALUE,
-					consumes = MediaType.APPLICATION_JSON_VALUE)
+	@RequestMapping(value = "/saveCard",
+					method = RequestMethod.POST)
 	@ResponseBody
-	public StripeChargeReceipt chargeCardAndSave(@RequestBody StripeCharge stripeCharge) throws Exception {
-		return this.stripeClient.chargePaymentCardAndSave(stripeCharge.getTokenId(),
-						stripeCharge.getAmountInclFees(),
-						stripeCharge.getAmountForGroup(),
-						stripeCharge.getUserId(),
-						stripeCharge.getGroupAccountId());
+	public ResponseEntity savePaymentCardDetails(@RequestBody StripeCharge customerDetails) {
+		return paymentService.savePaymentDetails(customerDetails);
 	}
 }
